move scanner bottom sheet out of combined scanner package
